requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# LangChain & LangGraph
langchain==0.1.0
langchain-openai==0.1.17
langchain-community==0.2.9
langchain-core==0.2.22
langchain-text-splitters==0.2.2
langgraph==0.0.20

# Vector Database
weaviate-client==4.4.0
langchain-weaviate==0.0.1

# OpenAI
openai==1.3.7

# Async Support
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# Data Processing
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2

# Web Scraping & APIs
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Text Processing
nltk==3.8.1
spacy==3.7.2

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Logging & Monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8